# actions_openapi.yaml
openapi: 3.1.0
info:
  title: Maverickframe HR Bridge
  version: 0.3.0
servers:
  - url: https://maverickframe-hh-bridge.onrender.com
paths:
...
  /sheets/save_candidate:
    post:
      operationId: saveCandidateToSheets
      summary: Save selected candidate to Google Sheets
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                date_added: { type: string }
                source: { type: string }
                vacancy: { type: string }
                candidate_name: { type: string }
                location: { type: string }
                experience: { type: string }
                skills: { type: string }
                gpt_score: { type: integer }
                status: { type: string }
                recruiter_comment: { type: string }
                resume_link: { type: string }
                suggested_reply: { type: string }
                gpt_summary: { type: string }
      responses:
        '200': { description: OK }